+++ internal/parser/parser_test.go
@@ -521,8 +521,12 @@ func TestScope(t *testing.T) {
 	expectParseError(t, "(function foo() { const foo = 0 })", "")
 
 	expectParseError(t, "var x; function x() {}", "")
+	expectParseError(t, "var x; function *x() {}", "<stdin>: error: \"x\" has already been declared\n")
+	expectParseError(t, "var x; async function x() {}", "<stdin>: error: \"x\" has already been declared\n")
 	expectParseError(t, "let x; function x() {}", "<stdin>: error: \"x\" has already been declared\n")
 	expectParseError(t, "function x() {} var x", "")
+	expectParseError(t, "function* x() {} var x", "<stdin>: error: \"x\" has already been declared\n")
+	expectParseError(t, "async function x() {} var x", "<stdin>: error: \"x\" has already been declared\n")
 	expectParseError(t, "function x() {} let x", "<stdin>: error: \"x\" has already been declared\n")
 	expectParseError(t, "function x() {} function x() {}", "")